pyproject.toml
[project]
name = "shprho-shoper"
description = "CLI to manage Multi-tenant deployments for ShprHO apps"
readme = "README.md"
license = "GPL-3.0-only"
requires-python = ">=3.7"
authors = [
    { name = "ShprHO Technologies Pvt Ltd", email = "developers@shopersolutions.com" },
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Natural Language :: English",
    "Operating System :: MacOS",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: User Interfaces",
    "Topic :: System :: Installation/Setup",
]
dependencies = [
    "Click>=7.0",
    "GitPython~=3.1.30",
    "honcho",
    "Jinja2~=3.0.3",
    "python-crontab~=2.6.0",
    "requests",
    "semantic-version~=2.8.2",
    "setuptools>40.9.0",
    "tomli;python_version<'3.11'",
]
dynamic = [
    "version",
]

[project.scripts]
shoper = "shoper.cli:cli"

[project.urls]
Changelog = "https://github.com/netmanthan/shoper/releases"
Documentation = "https://shprhoframework.com/docs/user/en/shoper"
Homepage = "https://shopersolutions.com/shoper"
Source = "https://github.com/netmanthan/shoper"

[build-system]
requires = [
    "hatchling>=1.6.0",
]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "shoper/__init__.py"

[tool.hatch.build.targets.sdist]
include = [
  "/shoper"
]

[tool.hatch.build.targets.wheel]
include = [
  "/shoper"
]
